+++ tests/testthat/test-read_parquet.R
@@ -25,8 +25,6 @@ test_that("read_parquet reads Parquet file without decryption", {
 # Test `read_parquet()` with decryption
 test_that("read_parquet reads encrypted Parquet file with decryption (legacy)", {
 
-  skip_on_cran()
-
   data <- tibble::tibble(a = 1:5, b = letters[1:5])
   temp_file <- tempfile(fileext = ".parquet")
 
@@ -56,8 +54,6 @@ test_that("read_parquet reads encrypted Parquet file with decryption (legacy)", 
 # Test `read_parquet()` with decryption
 test_that("read_parquet reads encrypted Parquet file with decryption", {
 
-  skip_on_cran()
-
   data <- tibble::tibble(a = 1:5, b = letters[1:5])
   temp_file <- tempfile(fileext = ".parquet")
 
@@ -81,3 +77,34 @@ test_that("read_parquet reads encrypted Parquet file with decryption", {
 })
 
 
+# Test `read_parquet()` with decryption
+test_that("read_parquet_as_db reads encrypted Parquet and returns a lazy table from DuckDB connection", {
+
+  data <- tibble::tibble(a = 1:5, b = letters[1:5])
+  temp_file <- tempfile(fileext = ".parquet")
+
+  mock_key <- 'rppqM5CuEqotys4wQq/g7xh6wpIjRozcAIbI9sagwKE='
+
+  # Write Parquet file with mock encryption
+  write_parquet(data, temp_file, encryption_key = mock_key)
+
+  # Read encrypted file with decryption
+  result <- read_parquet_as_db(temp_file, decryption_key = mock_key)
+
+  # Check if result is an Arrow table
+  expect_true(inherits(result, "tbl_duckdb_connection"))
+
+  result_df <- dplyr::collect(result)
+  # Check the content of the data
+  expect_equal(nrow(result_df), 5)
+  expect_equal(ncol(result_df), 2)
+
+  result_2 <- read_parquet_as_db(temp_file, decryption_key = mock_key, columns = 'a') |>
+    dplyr::collect()
+
+  expect_equal(nrow(result_2), 5)
+  expect_equal(ncol(result_2), 1)
+
+  unlink(temp_file, recursive = TRUE, force = TRUE)
+
+})